#227 moved buy&sell commands to module
